editComponent caching
fieldTag -> renamed to editTag for consistency
parseFieldArray -> no longer returns Field objects...just plain objects since field ViewModels inherit Field
