+++ docs-main.js
@@ -1,5 +1,9 @@
 import 'spectre.css/dist/spectre.css';
 import 'font-awesome/css/font-awesome.css';
 import 'animate.css/animate.css';
-import 'spectre-canjs/sp-form/demo/basic';
-import 'spectre-canjs/sp-form/demo/form';
+import 'spectre-canjs/sp-form/demo/basic/';
+import 'spectre-canjs/sp-form/demo/full/';
+import can from 'can-namespace';
+
+// for debugging
+window.can = can;
