+++ package.json
@@ -21,7 +21,7 @@
     "pretest": "npm run lint",
     "test": "jest",
     "build": "steal-tools",
-    "watch": "npm-watch",
+    "live": "steal-tools live-reload --main spectre-canjs/docs-main",
     "docs": "documentjs -d -f",
     "postdocs": "node docs-bundle",
     "prepublish-docs": "git checkout gh-pages && git merge master",
@@ -63,6 +63,7 @@
     "can-define": "^1.5.4",
     "can-event": "^3.7.6",
     "can-jquery": "^3.2.5",
+    "can-namespace": "^1.0.0",
     "can-stache": "^3.14.8",
     "can-stache-converters": "^3.3.0",
     "can-view-model": "^3.5.2",
