+++ sp-file-list/sp-file-list.js
@@ -4,7 +4,7 @@ import DefineList from 'can-define/list/';
 import './sp-file-list.less';
 import view from './sp-file-list.stache';
 
-const img = /(http(s?):)|([/|.|\w|\s])*\.(?:jpg|jpeg|gif|png)/;
+const img = new RegExp('(http(s?):)|([/|.|\\w|\\s])*\\.(?:jpg|jpeg|gif|png)', 'i');
 
 export const ViewModel = DefineMap.extend({
     files: DefineList,
