+++ sp-form/demo/basic/basic.html
@@ -1,0 +1,16 @@
+<script src="../../../node_modules/steal/steal.js" main="spectre-canjs/docs-main" id="demo-source">
+    // <script src="../../../docs/can-admin/steal.production.js" main="can-admin/index">
+    Promise.all([
+        steal.import('spectre-canjs/sp-form/demo/basic/basic'),
+        steal.import('live-reload')
+    ]).then(function([form, reload]){
+        form.default();
+        reload(function () {
+            const children = Array.from(document.body.children);
+            children.forEach((child) => {
+                child.remove();
+            });
+            form.default();
+        });
+    });
+</script>
