--- sp-form/demo/full.html
@@ -1,24 +1,0 @@
-<div id="demo-app"></div>
-
-<script type="text/stache" id="demo-html">
-  <div class="container">
-    <h1>Form Widget Demo</h1>
-    <p>Default form object values: </p>
-    <div class="json">{{{stringify (object)}}}</div>
-
-    <p>Dirty field values</p>
-    <div class="json">{{{stringify (dirtyObject)}}}</div>
-
-    <sp-form 
-      vm:object:bind="object" 
-      vm:dirtyObject:to="dirtyObject" 
-      vm:actions:from="actions"
-      on:vm:submit="onSubmit(scope.arguments)" 
-      vm:isSaving:bind="formSaving"
-      on:vm:cancel="onCancel"></sp-form>
-  </div>
-</script>
-<script src="../../node_modules/steal/steal.js" main="spectre-canjs/index" id="demo-source">
-// <script src="../../../docs/can-admin/steal.production.js" main="can-admin/index">
-  steal.import('spectre-canjs/sp-form/demo/form');
-</script>
