+++ util/field/Field.js
@@ -50,10 +50,10 @@ export const Field = DefineMap.extend('Field', {
      * The type of the form field to use when editing this field. These types
      * are defined in the `util/field.TEMPLATES` constant. This should be
      * omitted if a custom template is used.
-     * @property {String} util/field/Field.props.fieldTag fieldTag
+     * @property {String} util/field/Field.props.editTag editTag
      * @parent util/field/Field.props
      */
-    fieldTag: {
+    editTag: {
         type: 'string',
         value: 'sp-text-field'
     },
@@ -69,48 +69,13 @@ export const Field = DefineMap.extend('Field', {
      * @parent FieldInputMap.props
      */
     value: {
+        value: '',
         type: '*',
         set (val) {
             if (this.value !== val) {
                 this.dispatch('fieldchange', [this]);
             }
             return val;
-        }
-    },
-    /**
-     * The form field template to use when editing this field in the sp-form. This should be
-     * a template renderer. By default, this value is set to the
-     * template for the given `fieldTag` property.
-     *
-     * The default renderers are provided as a constant, and may be referenced
-     * by passing the `field.fieldTag` parameter. For instance, passing
-     * `fieldTag: 'select'` will set `editComponent` to the registered
-     * template for a `sp-select-field` component.
-     *
-     * Custom templates can be created to add various field types and functionality
-     * to the form widget.
-     *
-     * The custom templates will have the following useful properties in their scope:
-     *  - `this`: (alias `.` is the current `this` object) the field properties object
-     *  - `setField`: the function to call when the field changes
-     *  - `object`: the form object
-     *  - `validationErrors`: An object with keys referencing the field name, and a string referencing a validation error
-     *
-     * For example:
-     * @property {Renderer} util/field/Field.props.editComponent editComponent
-     * @parent util/field/Field.props
-     */
-    editComponent: {
-        type: '*',
-        get (renderer) {
-            if (typeof renderer === 'function') {
-                return renderer;
-            }
-            return stache(`<${this.fieldTag} 
-                properties:from="." 
-                value:bind="../dirtyObject[name]" 
-                error:bind="../validationErrors[name]"
-                on:fieldchange="checkField(scope.arguments)" />`);
         }
     },
     /**
