+++ util/field/parseFieldArray/parseFieldArray.test.js
@@ -14,6 +14,6 @@ test('parseFieldArray', () => {
 
     expect(fields.length).toEqual(2);
     fields.forEach((f) => {
-        expect(f instanceof Field).toBeTruthy();
+        expect(typeof f).toEqual('object');
     });
 });
